functions/requirements.txt
parse==1.19.0
yaspin==2.1.0
boto3==1.21.22
botocore==1.24.22
google-cloud-logging==3.0.0
pandas==1.3.5
matplotlib==3.5.1
seaborn==0.11.2
numpy==1.21.6
rich==12.0.0
notebook==6.4.8
